backend/requirements.txt
Flask==2.2.0
Werkzeug==2.2.3  # Version compatible avec Flask 2.2.0
flask-cors==3.0.10
cassandra-driver==3.28.0
pyspark==3.4.1
py4j==0.10.9.7
pywin32==306; platform_system == "Windows"
findspark==2.0.1
PyJWT==2.8.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
scikit-learn==1.3.0
joblib==1.3.2
